requirements.txt
Django==2.2.6
django-ckeditor==5.7.1
django-compressor==2.1
django-hitcount==1.3.1
django-uuslug==1.1.8
gunicorn==19.9.0
Pillow==6.2.0
pytz==2019.3
sorl-thumbnail==12.5.0